server/requirements.txt
fastapi
uvicorn[standard]
ffmpeg-python
numpy
git+https://github.com/m-bain/whisperx.git
openai-whisper
torch
